+++ src/main/java/underlay/packets/requests/SearchByNumIDRequest.java
@@ -3,11 +3,13 @@ package underlay.packets.requests;
 import underlay.packets.Request;
 import underlay.packets.RequestType;
 
+import java.math.BigInteger;
+
 public class SearchByNumIDRequest extends Request {
 
-    public final int targetNumID;
+    public final BigInteger targetNumID;
 
-    public SearchByNumIDRequest(int targetNumID) {
+    public SearchByNumIDRequest(BigInteger targetNumID) {
         super(RequestType.SearchByNumID);
         this.targetNumID = targetNumID;
     }
