+++ src/test/java/underlay/UnderlayTest.java
@@ -3,14 +3,13 @@ package underlay;
 import lookup.LookupTable;
 import lookup.LookupTableFactory;
 import middlelayer.MiddleLayer;
-import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
 import skipnode.SkipNode;
 import skipnode.SkipNodeInterface;
-import underlay.packets.RequestType;
 import underlay.packets.requests.*;
+
+import java.math.BigInteger;
 
 /**
  * This test creates two underlays on the machine at different ports and checks the
@@ -59,7 +58,7 @@ public class UnderlayTest {
         // Check search by name ID request.
         Assertions.assertNotNull(localUnderlay.sendMessage(remoteAddress, remotePort, new SearchByNameIDRequest("")));
         // Check search by numerical ID request.
-        Assertions.assertNotNull(localUnderlay.sendMessage(remoteAddress, remotePort, new SearchByNumIDRequest(0)));
+        Assertions.assertNotNull(localUnderlay.sendMessage(remoteAddress, remotePort, new SearchByNumIDRequest(BigInteger.valueOf(0))));
         // Check level-based search by name ID request.
         Assertions.assertNotNull(localUnderlay.sendMessage(remoteAddress, remotePort, new NameIDLevelSearchRequest(0, 0, "")));
         // Check left/right update requests.
